Fix IC Slot
修复16位 IC Slot 溢出的情景

Issue: #IA8ONB (ets_frontend)

+++ ecmascript/method.h
@@ -244,7 +244,7 @@ public:
         return MethodLiteral::GetMethodId(literalInfo);
     }
 
-    uint16_t GetSlotSize() const
+    uint32_t GetSlotSize() const
     {
         uint64_t literalInfo = GetLiteralInfo();
         return MethodLiteral::GetSlotSize(literalInfo);
